# docs/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ICoder API",
    "description": "API Documentation for ICoder Project",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/contests/{contestId}": {
      "get": {
        "tags": [
          "Contest Management"
        ],
        "summary": "Get contest details",
        "description": "Retrieves the details of a specific contest.",
        "operationId": "getContestDetails",
        "parameters": [
          {
            "name": "contestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ContestDetailsResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Contest Management"
        ],
        "summary": "Update an existing contest",
        "description": "Updates an existing contest with the provided details. Contest coordinators only can update contests.",
        "operationId": "updateContest",
        "parameters": [
          {
            "name": "contestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MessageResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Contest Management"
        ],
        "summary": "Delete a contest",
        "description": "Deletes an existing contest. Contest coordinators only can delete contests.",
        "operationId": "deleteContest",
        "parameters": [
          {
            "name": "contestId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/submissions": {
      "get": {
        "tags": [
          "Submission Management"
        ],
        "summary": "Get all submissions with filters (it could be used to reset too)",
        "description": "Retrieve paginated list of all submissions filtered by ser handle, online judge, problem code, and language",
        "operationId": "getAllSubmissions",
        "parameters": [
          {
            "name": "handle",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "online_judge",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "problem_code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "paging",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageSubmissionPageResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Submission Management"
        ],
        "summary": "Submit code",
        "description": "Submit code solution for a problem",
        "operationId": "submitCode",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionCreateResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/submissions/session": {
      "post": {
        "tags": [
          "Submission Management"
        ],
        "summary": "Add session id",
        "description": "add session id to submit with user account",
        "operationId": "addSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionSubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SessionSubmissionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/submissions/session/update": {
      "post": {
        "tags": [
          "Submission Management"
        ],
        "summary": "Update user session id",
        "description": "update user session id to submit with user account",
        "operationId": "updateSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionSubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SessionSubmissionResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/coding/editor/submissions": {
      "post": {
        "tags": [
          "Coding Editor"
        ],
        "summary": "Submit a single code snippet for execution",
        "description": "Accepts the source code, language ID, and standard input (stdin). A unique token is returned to poll for results (requires login)",
        "operationId": "submitCode_1",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/submissions/{submissionId}/toogle-openness": {
      "patch": {
        "tags": [
          "Submission Management"
        ],
        "summary": "Toggle submission openness",
        "description": "Toggle submission openness (hidden from others or not)",
        "operationId": "toggleOpenness",
        "parameters": [
          {
            "name": "submissionId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/submissions/{id}": {
      "get": {
        "tags": [
          "Submission Management"
        ],
        "summary": "Get submission by ID",
        "description": "Retrieve a specific submission by its ID",
        "operationId": "getSubmissionById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/submissions/session/{judgeType}": {
      "get": {
        "tags": [
          "Submission Management"
        ],
        "summary": "Get user session by judge type",
        "description": "get ",
        "operationId": "getSession",
        "parameters": [
          {
            "name": "judgeType",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SessionSubmissionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/submissions/languages/{oj}": {
      "get": {
        "tags": [
          "Submission Management"
        ],
        "summary": "Get languages for online judge",
        "description": "Retrieve list of languages supported by online judge",
        "operationId": "getLanguages",
        "parameters": [
          {
            "name": "oj",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LanguageOptionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/submissions/contests/{contest_id}": {
      "get": {
        "tags": [
          "Submission Management"
        ],
        "summary": "Get all submissions for specific contest with filters (it could be used to reset too)",
        "operationId": "getAllSubmissionsForContest",
        "parameters": [
          {
            "name": "contest_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "handle",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "result",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "language",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "problem_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "paging",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageContestSubmissionsResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/coding/editor/submissions/{token}": {
      "get": {
        "tags": [
          "Coding Editor"
        ],
        "summary": "Retrieve the execution result (Polling Endpoint)",
        "description": "Retrieves the current status and output of a submission. \n\n**Frontend Workflow:** \n1. If this endpoint returns **202 Accepted**, it means the code is still in queue or processing. The frontend should wait (e.g., 1-2 seconds) and poll again. \n2. If it returns **200 OK**, the execution is complete (regardless of whether the result is 'Accepted', 'Wrong Answer', or 'Error'). \n\n**Note:** Use the `status.id` to distinguish between 'Processing' (< 3) and 'Final' states (>= 3).",
        "operationId": "getResult",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/submissions/session/{id}": {
      "delete": {
        "tags": [
          "Submission Management"
        ],
        "summary": "Delete user session user",
        "operationId": "deleteSession",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessageResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "CreateContestRequest": {
        "required": [
          "length",
          "title"
        ],
        "type": "object",
        "properties": {
          "group_id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "begin_time": {
            "type": "string",
            "format": "date-time"
          },
          "length": {
            "pattern": "^\\d{1,3}:\\d{2}:\\d{2}$",
            "type": "string"
          },
          "contest_type": {
            "type": "string",
            "enum": [
              "CLASSICAL",
              "GROUP"
            ]
          },
          "contest_openness": {
            "type": "string",
            "enum": [
              "PUBLIC",
              "PROTECTED",
              "PRIVATE"
            ]
          },
          "password": {
            "type": "string"
          },
          "history_rank": {
            "type": "boolean"
          },
          "problem_set": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProblemSetRequest"
            }
          }
        }
      },
      "ProblemSetRequest": {
        "required": [
          "problem_alias"
        ],
        "type": "object",
        "properties": {
          "problem_id": {
            "type": "integer",
            "format": "int64"
          },
          "problem_alias": {
            "type": "string"
          },
          "problem_weight": {
            "type": "string"
          }
        }
      },
...
      "SubmissionCreateRequest": {
        "required": [
          "code",
          "language",
          "online_judge",
          "problem_code"
        ],
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "format": "int64"
          },
          "problem_code": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "language": {
            "type": "string"
          },
          "online_judge": {
            "type": "string",
            "enum": [
              "CODEFORCES",
              "GYM",
              "AT_CODER",
              "CSES"
            ]
          },
          "opened": {
            "type": "boolean"
          },
          "submission_method": {
            "type": "string",
            "enum": [
              "BOT",
              "SESSION"
            ]
          },
          "contest_id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "SubmissionCreateResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "remoteRunId": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "QUEUED",
              "SUBMITTING",
              "JUDGING",
              "COMPLETED",
              "FAILED"
            ]
          },
          "verdict": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "WRONG_ANSWER",
              "IN_QUEUE",
              "COMPILATION_ERROR",
              "RUNTIME_ERROR",
              "TIME_LIMIT_EXCEEDED",
              "FAILED",
              "RUNNING",
              "MEMORY_LIMIT_EXCEEDED",
              "FAILED_TO_SUBMIT"
            ]
          },
          "timeUsage": {
            "type": "integer",
            "format": "int32"
          },
          "memoryUsage": {
            "type": "integer",
            "format": "int32"
          },
          "problemCode": {
            "type": "string"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SessionSubmissionRequest": {
        "required": [
          "session_data"
        ],
        "type": "object",
        "properties": {
          "session_data": {
            "type": "string"
          },
          "online_judge": {
            "type": "string",
            "enum": [
              "CODEFORCES",
              "GYM",
              "AT_CODER",
              "CSES"
            ]
          }
        }
      },
      "SessionSubmissionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "SubmissionRequest": {
        "type": "object",
        "properties": {
          "source_code": {
            "type": "string"
          },
          "language_id": {
            "type": "integer",
            "format": "int32"
          },
          "stdin": {
            "type": "string"
          }
        }
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PageSubmissionPageResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SubmissionPageResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "PageableObject": {
        "type": "object",
        "properties": {
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "SubmissionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "remoteRunId": {
            "type": "string"
          },
          "onlineJudge": {
            "type": "string",
            "enum": [
              "CODEFORCES",
              "GYM",
              "AT_CODER",
              "CSES"
            ]
          },
          "language": {
            "type": "string"
          },
          "submittedAt": {
            "type": "string",
            "format": "date-time"
          },
          "memoryUsage": {
            "type": "string"
          },
          "timeUsage": {
            "type": "string"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "WRONG_ANSWER",
              "IN_QUEUE",
              "COMPILATION_ERROR",
              "RUNTIME_ERROR",
              "TIME_LIMIT_EXCEEDED",
              "FAILED",
              "RUNNING",
              "MEMORY_LIMIT_EXCEEDED",
              "FAILED_TO_SUBMIT"
            ]
          },
          "isOpen": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "CREATED",
              "QUEUED",
              "SUBMITTING",
              "JUDGING",
              "COMPLETED",
              "FAILED"
            ]
          },
          "problemCode": {
            "type": "string"
          },
          "userHandle": {
            "type": "string"
          }
        }
      },
      "LanguageOptionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          }
        }
      },
...
      "PageContestSubmissionsResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContestSubmissionsResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "ContestDetailsResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "begin_time": {
            "type": "string",
            "format": "date-time"
          },
          "end_time": {
            "type": "string",
            "format": "date-time"
          },
          "remaining_time": {
            "type": "object",
            "properties": {
              "seconds": {
                "type": "integer",
                "format": "int64"
              },
              "zero": {
                "type": "boolean"
              },
              "nano": {
                "type": "integer",
                "format": "int32"
              },
              "negative": {
                "type": "boolean"
              },
              "units": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "durationEstimated": {
                      "type": "boolean"
                    },
                    "timeBased": {
                      "type": "boolean"
                    },
                    "dateBased": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "length": {
            "type": "string"
          },
          "history_rank": {
            "type": "boolean"
          },
          "contest_status": {
            "type": "string",
            "enum": [
              "RUNNING",
              "SCHEDULED",
              "ENDED"
            ]
          },
          "contest_type": {
            "type": "string",
            "enum": [
              "CLASSICAL",
              "GROUP"
            ]
          },
          "owner_id": {
            "type": "integer",
            "format": "int64"
          },
          "owner_handle": {
            "type": "string"
          },
          "group_id": {
            "type": "integer",
            "format": "int64"
          },
          "group_name": {
            "type": "string"
          }
        }
      },
...
      "SubmissionResult": {
        "type": "object",
        "properties": {
          "stdout": {
            "type": "string"
          },
          "time": {
            "type": "number",
            "format": "float"
          },
          "memory": {
            "type": "integer",
            "format": "int32"
          },
          "stderr": {
            "type": "string"
          },
          "token": {
            "type": "string"
          },
          "compile_output": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/Status"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}